--- public/index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <base href="/">
-    <link href="https://fonts.googleapis.com/css?family=Lato" rel="stylesheet">
-    <meta charset="utf-8">
-    <title>Sylcat | Parkour social network</title>
-  </head>
-  <body>
-    <div id="sylcat"></div>
-  </body>
-</html>
